backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
tensorflow
numpy
faiss-cpu